+++ src/components/Box/RowBox.js
@@ -2,7 +2,12 @@ import React from 'react';
 import { Box } from '@material-ui/core';
 import PropTypes from 'prop-types';
 
-export default function RowBox({ fluid, reverse, direction: _direction, ...props }) {
+export default function RowBox({
+  fluid,
+  reverse,
+  direction: _direction,
+  ...props
+}) {
   let direction = _direction;
 
   if (reverse) {
@@ -20,11 +25,11 @@ export default function RowBox({ fluid, reverse, direction: _direction, ...props
 }
 
 RowBox.defaultProps = {
-  direction: 'row'
-}
+  direction: 'row',
+};
 
 RowBox.propTypes = {
   fluid: PropTypes.bool,
   reverse: PropTypes.bool,
-  direction: PropTypes.oneOf(['row', 'column'])
-}
+  direction: PropTypes.oneOf(['row', 'column']),
+};