+++ src/components/List/SoundListHeader/SoundListHeader.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { Typography } from '@material-ui/core';
 import { Category } from '@material-ui/icons';
 
-import { RowBox, SecondaryButton } from '../../../../../../components';
-import { useOrientation } from '../../../../../../hooks';
+import { RowBox, SecondaryButton } from '../../.';
+import { useOrientation } from '../../../hooks';
 
 export default function SoundListHeader({ onCategoriesVisibilityChange }) {
   const { portrait } = useOrientation();
 
   return (
-    <RowBox>
+    <RowBox justifyContent="space-between">
       <Typography variant="h4" color="textPrimary" gutterBottom>
         Browse sounds
       </Typography>