+++ src/components/List/index.js
@@ -1,0 +1,3 @@
+export { default as SoundPlaylist } from './SoundPlaylist';
+export { default as SoundCategoriesList } from './SoundCategoriesList';
+export { default as SoundListHeader } from './SoundListHeader';