+++ src/components/Sounds/SoundCardList/SoundCardList.js
@@ -3,17 +3,21 @@ import { useSelector, useDispatch } from 'react-redux';
 import { Grid } from '@material-ui/core';
 
 import SoundCard from '../SoundCard';
-import { getPlaylistItems, addToPlaylist, removeFromPlaylist } from '../../../redux/playlist';
+import {
+  getPlaylistItems,
+  addToPlaylist,
+  removeFromPlaylist,
+} from '../../../redux/playlist';
 
 export default function SoundCardList({ data }) {
   const playlistItems = useSelector(getPlaylistItems);
   const dispatch = useDispatch();
 
-  const checkItem = id => playlistItems.find(element => element.id === id);
+  const checkItem = (id) => playlistItems.find((element) => element.id === id);
 
   return (
-    <Grid container spacing={2} style={{ marginTop: '1rem' }}>
-      {data.map(item => (
+    <Grid container spacing={2} style={{ marginTop: '1rem', marginBottom: 0 }}>
+      {data.map((item) => (
         <Grid item lg={4} md={6} sm={12} key={item.id}>
           <SoundCard
             name={item.name}