+++ src/components/index.js
@@ -10,3 +10,4 @@ export * from './MediaPlayer';
 export { default as Loading } from './Loading';
 export * from './Box';
 export { default as SidebarMenu } from './SidebarMenu';
+export * from './List';