+++ src/pages/MainPage/pages/BrowsePage/BrowsePage.js
@@ -1,17 +1,14 @@
 import React, { useState, useEffect } from 'react';
-import { Grid, Paper, Collapse } from '@material-ui/core';
 
-import { Loading, ColumnBox } from '../../../../components';
 import { OrientationRoute } from '../../../../containers';
-import SoundList from './components/SoundList';
-import SoundCategoriesList from './components/SoundCategoriesList';
-import SoundListHeader from './components/SoundListHeader';
 import { getSets } from '../../../../api/sets';
+
+import BrowseLandscape from './pages/BrowseLandscape';
+import BrowsePortrait from './pages/BrowsePortrait';
 
 export default function BrowsePage() {
   const [loading, setLoading] = useState(true);
   const [sets, setSets] = useState([]);
-  const [showCategories, setShowCategories] = useState(false);
 
   useEffect(() => {
     (async () => {
@@ -27,52 +24,8 @@ export default function BrowsePage() {
 
   return (
     <OrientationRoute
-      landscape={
-        <Grid container direction="row" style={{ flex: 1 }}>
-          <Grid item md={2}>
-            <Paper style={{ minHeight: '100%' }}>
-              <SoundCategoriesList />
-            </Paper>
-          </Grid>
-          <Grid item md={10}>
-            {loading ? (
-              <Loading />
-            ) : (
-              <SoundList
-                data={sets}
-                header={
-                  <SoundListHeader
-                    onCategoriesVisibilityChange={() =>
-                      setShowCategories(!showCategories)
-                    }
-                  />
-                }
-              />
-            )}
-          </Grid>
-        </Grid>
-      }
-      portrait={
-        <ColumnBox>
-          <Collapse in={showCategories}>
-            <SoundCategoriesList />
-          </Collapse>
-          {loading ? (
-            <Loading />
-          ) : (
-            <SoundList
-              data={sets}
-              header={
-                <SoundListHeader
-                  onCategoriesVisibilityChange={() =>
-                    setShowCategories(!showCategories)
-                  }
-                />
-              }
-            />
-          )}
-        </ColumnBox>
-      }
+      landscape={<BrowseLandscape loading={loading} sounds={sets} />}
+      portrait={<BrowsePortrait loading={loading} sounds={sets} />}
     />
   );
 }