+++ src/pages/MainPage/styled/MainContainer.js
@@ -5,9 +5,9 @@ const MainContainer = styled(Box)(({ theme }) => ({
   padding: 0,
   margin: 0,
   backgroundColor: `${theme.palette.background.default}`,
-  height: '100%',
+  minHeight: '100vh',
   display: 'flex',
-  flexDirection: 'column'
+  flexDirection: 'column',
 }));
 
 export default MainContainer;