+++ src/scripts/installSound/constants.js
@@ -1,9 +1,15 @@
 const SOUNDS_PATH = './public/data/sounds';
-const ALLOWED_EXTENSIONS = ['.wav', 'mp3'];
-const ALLOWED_TYPES = ['sound', 'parallel container', 'sequence container', 'random container', 'sound fx'];
+const ALLOWED_EXTENSIONS = ['.wav', '.mp3'];
+const ALLOWED_TYPES = [
+  'sound',
+  'parallel container',
+  'sequence container',
+  'random container',
+  'sound fx',
+];
 
 module.exports = {
   SOUNDS_PATH,
   ALLOWED_EXTENSIONS,
-  ALLOWED_TYPES
-}
+  ALLOWED_TYPES,
+};